Move utils class to utils folder
