+++ common/app/com/commercetools/sunrise/common/utils/ReflectionUtils.java
@@ -1,4 +1,4 @@
-package com.commercetools.sunrise.common.reverserouter;
+package com.commercetools.sunrise.common.utils;
 
 public final class ReflectionUtils {
